Remove callback proc as we can rely on Kernal.loop
The main menu is already being called infinitely until exit is called
so main menu will be called after invalid input.

We have to be careful to return from an invalid input so we the main
menu can be looped.

That is why the invalid input call from
choose_person_type was moved to the outer method

The other option is to raise and Invalid input error to and then rescue
it in the loop